pyproject.toml
[project]
name = "reflection-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "langchain>=1.2.0",
    "langchain-openai>=1.1.4",
    "langgraph>=1.0.5",
    "python-dotenv>=1.2.1",
    "grandalf>=0.8",
]
